% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend_helpers.R
\name{apply_measure}
\alias{apply_measure}
\title{Apply a measure to the current state of area assignments}
\usage{
apply_measure(areas, measure)
}
\arguments{
\item{areas}{one-row data frame with each column representing a partial area
of the total block area, in m2, as e.g. returned by
\code{\link{rabimo_block_to_partial_areas_m2}}}

\item{measure}{list with elements "name" (name of measurement) and
area (area in m2 that is assigned to the measure)}
}
\description{
Apply a measure to the current state of area assignments
}
