% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend_helpers.R
\name{get_available_m2}
\alias{get_available_m2}
\title{Get available area for each measure, based on current "state"}
\usage{
get_available_m2(areas)
}
\arguments{
\item{areas}{areas in m2, as returned by rabimo_block_to_partial_areas_m2}
}
\description{
Get available area for each measure, based on current "state"
}
