% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend_helpers.R
\name{get_measure_names}
\alias{get_measure_names}
\title{Get names of measures supported by kwb.smartwater}
\usage{
get_measure_names()
}
\description{
Get names of measures supported by kwb.smartwater
}
