% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend_helpers.R
\name{get_test_block}
\alias{get_test_block}
\title{Get one block (columns as expected by kwb.rabimo) for testing}
\usage{
get_test_block()
}
\description{
Get one block (columns as expected by kwb.rabimo) for testing
}
