% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend_helpers.R
\name{get_test_blocks}
\alias{get_test_blocks}
\title{Get one block (columns as expected by kwb.rabimo) for testing}
\usage{
get_test_blocks(codes = c("1100541241000000", "1400761421000000"))
}
\arguments{
\item{codes}{codes of the blocks to be selected from the Berlin dataset}
}
\description{
Get one block (columns as expected by kwb.rabimo) for testing
}
