% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_class_functions.R
\name{get_veg_class_increment}
\alias{get_veg_class_increment}
\title{Vegetation Class Increment Caused by Planting Trees}
\usage{
get_veg_class_increment(tree_measure_volume, unsealed_area_m2, m = 4.7)
}
\arguments{
\item{tree_measure_volume}{assumed green volume of trees in m3.}

\item{unsealed_area_m2}{unpaved area (= total_area * (1 - roof - pvd)) in m2}

\item{m}{slope of the linear relation between normalised vegetation volume
per unsealed area (vegnorm) and veg_scaled. Default: 4.7, calculated from
block with highest vegetation class in Berlin (code = 0000000012002198):
m = veg_scaled/vegnorm = 118.1306/24.92435}
}
\description{
Vegetation Class Increment Caused by Planting Trees
}
