% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effect_of_disconnect.R
\name{plot_effect_of_disconnect}
\alias{plot_effect_of_disconnect}
\title{Plot Effect of Disconnecting Surfaces from Sewer System}
\usage{
plot_effect_of_disconnect(
  surface_reduction,
  type,
  output_dir = NULL,
  width_factor = 10/6.789581
)
}
\arguments{
\item{surface_reduction}{reduction of connected surface in percent}

\item{type}{one of "critical_hours", "unpleasant_hours", "critical_events",
"negative_deviation"}

\item{output_dir}{if \code{NULL} (the default), the plot goes into the active
device, otherwise the plot is written to a png file within
\code{output_dir} (must be an existing directory)}

\item{width_factor}{width factor. Default: 10/6.789581}
}
\value{
path to created file (invisibly) if \code{output_dir} is not
\code{NULL}, otherwise \code{NULL} (invisibly)
}
\description{
Plot Effect of Disconnecting Surfaces from Sewer System
}
