% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend_helpers.R
\name{rabimo_block_to_partial_areas_m2}
\alias{rabimo_block_to_partial_areas_m2}
\title{Convert R-Abimo-block to partial areas given in m2}
\usage{
rabimo_block_to_partial_areas_m2(block)
}
\arguments{
\item{block}{R-Abimo block, as e.g. returned by \code{\link{get_test_blocks}}}
}
\description{
Convert R-Abimo-block to partial areas given in m2
}
