% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend_helpers.R
\name{test_plumber_api}
\alias{test_plumber_api}
\title{Test the plumber API}
\usage{
test_plumber_api(catch_errors = FALSE)
}
\arguments{
\item{catch_errors}{whether or not to catch errors. If TRUE, errors are
caught and the result object is a list with elements "data" (containing the
function result or NULL in case of an error) and "error" (containing the
error message in case of an error and NULL otherwise)}
}
\description{
Test the plumber API
}
