% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend_helpers.R
\name{to_plumber_response}
\alias{to_plumber_response}
\title{Convert function result (may inherit from "try-error") to a response}
\usage{
to_plumber_response(result)
}
\arguments{
\item{result}{result of calling a function within try()}
}
\description{
Convert function result (may inherit from "try-error") to a response
}
