% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_water_balance.R
\name{calculate_water_balance}
\alias{calculate_water_balance}
\title{Calculate Water Balance for Given Block Areas and Measures}
\usage{
calculate_water_balance(blocks, measures, convert_types = FALSE)
}
\arguments{
\item{blocks}{data.frame as returned by \code{\link{get_test_blocks}}}

\item{measures}{data.frame containing information about the planned measures
in m2. Each row refers to a block area, linked by the text field
\code{code}. There is one numeric field per measure. The names of the
measure-related fields must correspond to the names returned by
\code{\link{get_measure_names}}.}
}
\description{
Calculate Water Balance for Given Block Areas and Measures
}
