% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend_helpers.R
\name{get_measure_info}
\alias{get_measure_info}
\title{Get info on the measures supported by kwb.smartwater}
\usage{
get_measure_info(type = character(0), field_name_only = FALSE)
}
\arguments{
\item{type}{optional. Vector of character indicating the measure types
("green_roof", "pavement", "trees", "infiltration", "retention") for which
to filter the output.}

\item{field_name_only}{optional. Logical of length one indicating whether or
not to return only the "field_name" instead of all info fields per measure}
}
\description{
Get information on the rainwater management measures supported by
kwb.smartwater
}
